% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex_citation_session.R
\name{bibtex_citation_session}
\alias{bibtex_citation_session}
\title{Bibtex citation of Session}
\usage{
bibtex_citation_session(output_file = "R_attached_packages_citations.txt")
}
\arguments{
\item{output_file}{Character. Name/Location of the TXT file where Bibtex citations of currently loaded packages will be saved.}
}
\value{
TXT-File with Bibtex citations of all packages currently loaded in the Session, ready to be pasted into Latex. Citation keys are automatically generated, using the package name.
}
\description{
Bibtex citation of Session
}
\examples{
bibtex_citation_session(tempfile(fileext = ".txt"))
}
