% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_sh.R
\name{config_sh}
\alias{config_sh}
\title{Create bash config file for the canopy height model}
\usage{
config_sh(directory, tile, year)
}
\arguments{
\item{directory}{Parent directory of the model deployment}

\item{tile}{Tile to be predicted}

\item{year}{Year as character}
}
\value{
bash script file
}
\description{
Create bash config file for the canopy height model
}
\examples{
config_sh("C:/Users/esanc/Documents/", "T32ULC", "2020")
}
