% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param_df.R
\name{create_param_df}
\alias{create_param_df}
\title{Create a data frame of parameter combinations}
\usage{
create_param_df(
  tiles,
  bands,
  increments,
  decrease,
  year,
  base_folder,
  worldcover = "2020"
)
}
\arguments{
\item{tiles}{Vector of Senitnel-2 tile names. Ex: c("T32TMT")}

\item{bands}{Vector of Sentinel Bands to be manipulated}

\item{increments}{Integer vector. Increments by which the Bands should be increased or decreased.}

\item{decrease}{Character. If "True" increments are decreased, if "False" increment is increased.}

\item{year}{Character. Year of the Senitnel-2 imagery.}

\item{base_folder}{Character. Directory that is used for the analysis.}

\item{worldcover}{Character. Default is "2020".}
}
\value{
A data frame with all possible combinations of variables.
}
\description{
Create a data frame of parameter combinations
}
\examples{
variables <- create_param_df(tiles = c("T32TMT"),
                                       bands = c("B02","B08"),
                                       increments = c(0.05, 0.1, 0.15, 0.2),
                                       decrease = c("False"),
                                       year = "2020",
                                       base_folder = "./home/emilio/wip/global-canopy-height-model")
}
