% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_raster_with_s2.R
\name{crop_raster_with_s2}
\alias{crop_raster_with_s2}
\title{Title}
\usage{
crop_raster_with_s2(
  raster_dir,
  s2_dir,
  out_dir,
  out_basename,
  filetype = "tif"
)
}
\arguments{
\item{raster_dir}{Character. Filepath to directory with the rasters to be clipped. Expects name to contain S2-Tilename.}

\item{s2_dir}{Character. Filepath to directorywith the S2-scenes (just one jp200 or tif, not whole folder.) Must match the number of inout rasters and be named similar, as they are processed by order.}

\item{out_dir}{Character. Output directory for the final files.}

\item{out_basename}{Character. Basename to which the Sentinel tile name is appended.}

\item{filetype}{Character. Default "tif".}
}
\value{
Rasters as tif cropped to extend of S2-images.
}
\description{
Title
}
