% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sorter.R
\name{file_sorter}
\alias{file_sorter}
\title{Sort files within a directory into folders by group}
\usage{
file_sorter(path, by = "type")
}
\arguments{
\item{path}{character. A file path to the folder containing the files that should be organised. Accepts also zip-folders.}

\item{by}{character. Default "type". Parameter by which the folders should be organised. Valid values c("type", "date").}
}
\value{
Returns the folder "files_sorted" including all the files sorted into folders by their file-type. A copy summary is displayed after execution.
}
\description{
Sort files within a directory into folders by group
}
