% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecopy.batch.R
\name{filecopy.batch}
\alias{filecopy.batch}
\title{Batch extraction by file type sorted in folders}
\usage{
filecopy.batch(path, by = "type")
}
\arguments{
\item{path}{A file path to the desired folder - needs to end with "/"}

\item{by}{default "type" | no other options}
}
\value{
A set of folders for every filetype in the starting directory including subfolders
}
\description{
Batch extraction by file type sorted in folders
}
