% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_point_grid.r
\name{global_point_grid}
\alias{global_point_grid}
\title{Global Point grid}
\usage{
global_point_grid(
  lon_grid = 20,
  lat_grid = 15,
  lon_border = c(-180, 180),
  lat_border = c(-60, 60),
  coast_buffer = -0.05
)
}
\arguments{
\item{lon_grid}{Numeric. Number of Longitude points spanning the grid.}

\item{lat_grid}{Numeric. Number of Latitude points spanning the grid.}

\item{lon_border}{Vector, Numeric. Outer Longitude borders of the desired grid.}

\item{lat_border}{Vector, Numeric. Outer Latitude borders of the desired grid.}

\item{coast_buffer}{Numeric. Buffer around Coastlines.}
}
\value{
Character vector of Latitude and Longitude pairs, spanning a global point grid over the landmasses of earth.
}
\description{
Global Point grid
}
\examples{
global_point_grid(lon_grid = 20, lat_grid = 15, lon_border = c(-180,180),
                            lat_border = c(-60,60), coast_buffer = -0.05)

}
