% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANDELION.R
\name{nonnullPropEst}
\alias{nonnullPropEst}
\title{Estimate the Non-null Proportion}
\usage{
nonnullPropEst(x, u, sigma)
}
\arguments{
\item{x}{Numeric vector.}

\item{u}{Null mean.}

\item{sigma}{Null standard deviation.}
}
\value{
Estimated non-null proportion.
}
\description{
Internal helper used by DANDELION.
}
\keyword{internal}
