% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANDELION.R
\name{print.dandelion_pairs}
\alias{print.dandelion_pairs}
\title{Print DANDELION pair results}
\usage{
\method{print}{dandelion_pairs}(x, ...)
}
\arguments{
\item{x}{A \code{dandelion_pairs} object returned by \code{calc_pair.gene()} or
\code{calc_pair.snp()}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print DANDELION pair results
}
