% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANDELION.R
\name{print.dandelion_result}
\alias{print.dandelion_result}
\title{Print a DANDELION result object}
\usage{
\method{print}{dandelion_result}(x, ...)
}
\arguments{
\item{x}{A \code{dandelion_result} object returned by \code{med_gene()}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a DANDELION result object
}
