% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANDELION.R
\name{calc_pair.gene}
\alias{calc_pair.gene}
\title{Cleanse the pairs and disease proximal genes identified by DENDELION and map the disease proximal genes to WES significant genes (FOR genes 1 is gene.)}
\usage{
calc_pair.gene(mat.sig, mat.p, p.wes, gene1, ref.table.keep, eta.wgs = 1e-05)
}
\arguments{
\item{mat.sig}{matrix encoding significant pair, output by med_gene().}

\item{mat.p}{matrix encoding DENDELION P-values, output by med_gene()}

\item{p.wes}{a vector of length K. p value for gene 2 -> trait. Both p.trans and p.wes should contain the gene name.}

\item{gene1}{gene 1 names used for analysis, output by med_gene()}

\item{ref.table.keep}{reference data for position information for 'lincRNA' and 'protein_coding' which are not located in 'chrX', 'chrY' and 'chrM'.}

\item{eta.wgs}{the significant threshold for WES gene}
}
\value{
a list includes identified pairs, combined pairs (loci based), identified disease proximal genes (also WES significant), and identified disease proximal genes (NOT WES significant)
}
\description{
Organize the output results of med_gene() and merge nearby genes into a single locus.
}
