% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANDELION.R
\name{calc_pair.snp}
\alias{calc_pair.snp}
\title{Cleanse the pairs and disease proximal genes identified by DANDELION and map the disease proximal genes to WES significant genes (FOR genes 1 is SNPs)}
\usage{
calc_pair.snp(
  mat.sig,
  mat.p,
  p.wes,
  gene1,
  uniq_snp,
  ref.table.keep,
  eta.wgs = 1e-05,
  GRCh = "37"
)
}
\arguments{
\item{mat.sig}{matrix encoding significant pair, output by med_gene().}

\item{mat.p}{matrix encoding DANDELION P-values, output by med_gene()}

\item{p.wes}{a vector of length K. p value for gene 2 -> trait. Both p.trans and p.wes should contain the gene name.}

\item{gene1}{candidate SNPs names used for analysis, output by med_gene()}

\item{uniq_snp}{matrix should contains columns about the snp (named as 'SNP') and the corresponding significant cis gene name (named as 'GeneSymbol')}

\item{ref.table.keep}{reference data for position information for 'lincRNA' and 'protein_coding' which are not located in 'chrX', 'chrY' and 'chrM'.}

\item{eta.wgs}{the significant threshold for WES gene}
}
\value{
a list includes identified pairs, combined pairs (loci based), identified disease proximal genes (also WES significant), and identified disease proximal genes (NOT WES significant)
}
\description{
Organize the output results of med_gene() and merge nearby genes into a single locus.
}
