% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANDELION.R
\name{med_gene}
\alias{med_gene}
\title{Apply DANDELION to identify candidate disease proximal genes.}
\usage{
med_gene(
  p.trans,
  p.wes,
  ref.table,
  gene1.list,
  target.fdr = 0.1,
  dist = 5e+06,
  gene1.type = c("SNP", "Gene"),
  SNP.ref = NULL
)
}
\arguments{
\item{p.trans}{A matrix of p-values for distal-> proximal (gene1 -> gene2) associations.
The matrix has dimensions M (disease proximal genes, gene2) x N (disease distal genes, gene1).
Each entry represents the p-value for a potential trans effect from a gene1 to a gene2.
Both row names (gene2) and column names (gene1) should be gene symbols.}

\item{p.wes}{A named numeric vector containing p-values for associations between disease proximal genes (gene2) and the trait.
The names should correspond to core gene identifiers (gene2), and must match the row names of p.trans.}

\item{ref.table}{reference data for position information.}

\item{gene1.list}{A character vector of candidate disease distal gene names (gene1),
which must be a subset of the column names of p.trans.}

\item{target.fdr}{the pre-defined FDR level in the BH step, the default value is 0.1.}

\item{dist}{the distant between the gene and trans-genes, the default value is 5e6}

\item{gene1.type}{specify the exposure type. The expousre can be SNP, or Gene.}

\item{SNP.ref}{reference data for position information if the exposure is 'SNP'
should contain three columns:
1. SNP: rsid
2. SNPPos: position of SNP
3. SNPChr: '1'}
}
\value{
A list of candidate genes 1, identified pair matrix, and DANDELION's p-values
}
\description{
Apply DANDELION using p-values from WES and eQTLGen data to generate DANDELION's p-values and identified pairs. For code clarity, disease distal genes are abbreviated as 'gene1', and disease proximal genes as 'gene2' within the function.
}
