% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_manipulate.R
\name{s2_manipulate}
\alias{s2_manipulate}
\title{Manipulate single Bands of Sentinel 2 images by a percentage}
\usage{
s2_manipulate(zip_path, band_name, manipulation = "increase", increment = 0.1)
}
\arguments{
\item{zip_path}{Path to the zip file to be manipulated}

\item{band_name}{Name of the S2 Band to be manipulated (official names)}

\item{manipulation}{Increase or decrease}

\item{increment}{Ammount of change applied to the band}
}
\value{
Zip folder in the original structure with manipulated Band within a folder structure
}
\description{
Manipulate single Bands of Sentinel 2 images by a percentage
}
