% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldcover_adjust.R
\name{worldcover_adjust}
\alias{worldcover_adjust}
\title{Adjust worldcover image to the Structure of a Sentinel2 tile.}
\usage{
worldcover_adjust(wcover_tiles, wc_tile_status)
}
\arguments{
\item{wcover_tiles}{Character. File path to the worldcover image to be adjusted. CRS, extent and resolution gets modified if needed.}

\item{wc_tile_status}{Data frame with Tile names and status of the Worldcover image Tile (if it already was processed). Included for looping to not perform the same calculation again and again.}
}
\value{
Creates a new tif that is adjusted to the Sentinel2 Format and overwrites the original input.
}
\description{
Adjust worldcover image to the Structure of a Sentinel2 tile.
}
